🚚 Rename: 포스트 페이지 파일 구조 변경 (#227)
